Adam and me: Fixed: new log, update log, delete log Started: Create user

+++ app/app.py
@@ -10,7 +10,6 @@ from . import db
 app = Flask(__name__)
 app.secret_key = b'f01790380dc11025fcc4c8008d127a3b9da5f20199ce85efbd4bab36f68bd43d'
 
-
 app.config.from_mapping(
     DATABASE=os.path.join(app.instance_path, 'tourdeflask.sqlite'),
 )
@@ -58,7 +57,8 @@ def dev_form():  # put application's code here
     devs = db.select_all_users()
     log = []
     langs = db.list_langs()
-    return render_template('dev_list.html', devs=devs, log=log, langs=langs)
+    user_id = session['admin']
+    return render_template('dev_list.html', devs=devs, log=log, langs=langs, user_id=user_id)
 
 
 @app.route('/edit_log/<int:log_id>')
@@ -74,25 +74,36 @@ def logout():
     # remove the username from the session if it's there
     session.pop('user_id', None)
     session.pop('username', None)
-    session.pop('bool_admin', None)
+    session.pop('admin', None)
     return redirect('/')
 
 
 @app.route('/delete_dev/<int:developer_id>')
 def dev_delete(developer_id):
-    db.delete_dev(developer_id)
-    return redirect('/devs')
+    if session['admin']:
+        db.delete_dev(developer_id)
+        return redirect('/devs')
+    abort(400, 'Error - Invalid permission')
 
 
 @app.route('/create_dev', methods=['POST', 'GET'])
 def dev_insert():  # put application's code here
-    if request.method == 'POST':
-        new_name = request.form.get('new_name', type=str)
-        if new_name != '':
-            result = db.insert_dev(new_name)
-            if result == 'Error':
-                abort(400, 'Name already exists.')
-    return redirect('/devs')
+    if session['admin']:
+        if request.method == 'POST':
+            fname = request.form.get('fname', type=str).strip()
+            lname = request.form.get('lname', type=str).strip()
+            username = request.form.get('username', type=str).strip()
+            mail = request.form.get('mail', type=str).strip()
+            password = request.form.get('password', type=str)
+            bool_admin = request.form.get('bool_admin', type=int)
+            if all([fname != "", lname != "", username != "", mail != "", password != "", ]):
+                result = db.insert_dev(fname, lname, username, mail, password, bool_admin)
+                if "Error" in result:
+                    abort(400, result)
+            else:
+                abort(400, "Error - something is missing")
+        return redirect('/devs')
+    return redirect('/')
 
 
 @app.route('/edit_dev/<int:developer_id>', methods=['POST', 'GET'])
@@ -105,65 +116,41 @@ def dev_edit(developer_id):  # put application's code here
     return redirect('/devs')
 
 
-@app.route('/new/All', methods=['POST', 'GET'])
+@app.route('/new_log', methods=['POST', 'GET'])
 def log_form():  # put application's code here
     if request.method == 'POST':
-        name = request.form.get('dev', type=str)
+        dev_id = session["user_id"]
         work_date = request.form.get('work_date', type=str)
         lang = request.form.get('lang', type=str)
         duration = request.form.get('duration', type=int)
         rating = request.form.get('rating', type=int)
         note = request.form.get('note', type=str)
-        result = db.insert_log(name, work_date, lang, duration, rating, note)
+        result = db.insert_log(dev_id, work_date, lang, duration, rating, note)
         if result == 'Error':
             abort(400, 'Developer does not exist.')
     return redirect('/')
-
-
-@app.route('/dev/new/<int:dev_id>', methods=['POST', 'GET'])
-def dev_log_form(dev_id):  # put application's code here
-    if request.method == 'POST':
-        name = request.form.get('dev', type=str)
-        work_date = request.form.get('work_date', type=str)
-        lang = request.form.get('lang', type=str)
-        duration = request.form.get('duration', type=int)
-        rating = request.form.get('rating', type=int)
-        note = request.form.get('note', type=str)
-        result = db.insert_log(name, work_date, lang, duration, rating, note)
-        if result == 'Error':
-            abort(400, 'Developer does not exist.')
-    return redirect('/dev/' + str(dev_id))
 
 
 @app.route('/delete_log/<int:log_id>')
 def log_delete(log_id):
-    db.delete_log(log_id)
+    result = db.delete_log(log_id, session['user_id'])
+    if "Error" in result:
+        abort(400, result)
     return redirect('/')
-
-
-@app.route('/dev/delete_log/<int:log_id>')
-def dev_log_delete(log_id):
-    log = db.select_one_log(log_id)
-    dev_id = log['developer_id']
-    db.delete_log(log_id)
-    return redirect('/dev/' + str(dev_id))
 
 
 @app.route('/edit/<int:log_id>', methods=['POST', 'GET'])
 def log_update(log_id):
-    dev_id = ''
     if request.method == 'POST':
-        name = request.form.get('dev', type=str)
         work_date = request.form.get('work_date', type=str)
         lang = request.form.get('lang', type=str)
         duration = request.form.get('duration', type=int)
         rating = request.form.get('rating', type=int)
         note = request.form.get('note', type=str)
-        result = db.update_log(log_id, name, work_date, lang, duration, rating, note)
-        dev_id = str(db.dev_name_to_id(name))
-        if result == 'Error':
-            abort(400, 'Developer or log does not exist.')
-    return redirect('/dev/' + dev_id)
+        result = db.update_log(log_id, session['user_id'], work_date, lang, duration, rating, note)
+        if "Error" in result:
+            abort(400, result)
+    return redirect('/')
 
 
 if __name__ == '__main__':
